* commit '0aeedee9825dcf342c1fe156a7adbb8c4919d4ec':
  foutje in events

+++ ChipDistributionGame/Shared/src/main/java/nl/han/asdb/shared/interfaces/eventmanager/events/game/SellChipsEvent.java
@@ -51,7 +51,7 @@ public class SellChipsEvent extends Event {
         return amount;
     }
 
-    public void setAmount(short amount) {
+    public void setAmount(double amount) {
         this.amount = amount;
     }
